% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixRaschTools.R
\docType{data}
\name{fourclass_ex}
\alias{fourclass_ex}
\title{mixRaschTools Sample Four Class Model -----------------------------------------------------------}
\format{This four class mixed Rasch model was created with  the use of the mixRasch program (Willse,
2014) using a subset of the 2012 PISA US  math achievement data (OECD, 2012).  This model was created using  the "Math Self Efficacy" scale included in the
in the contextual assessment included in that year.  The dataset used for the models was created using students from the United States who had completed all
math questions contained in the PM7A booklet.  The final dataset contained 12 dichotomously scored
math achievement items creating a sample of 1229 students.  These models contain the following parameters:
"fourclass_ex" contains the ability estimates, item parameters, and standard errors for a four class Rasch model fitted using the mixRasch program.
"threeclass_ex" contains the ability estimates, item parameters, and standard errors for a three class Rasch model fitted using the mixRasch program.}
\source{
{http://www.oecd.org/pisa/pisaproducts/pisa2012database-downloadabledata.htm}
}
\usage{
fourclass_ex
}
\description{
Object containing model information for a four class model.
}
\references{
Organisation for Economic Co-operation and Development., & Programme for International
      Student Assessment. (2012a).  Paris: OECD.
Willse, J. T. (2014).  Package mixRasch.  R package version 1.1.
}
\keyword{datasets}
