% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixRaschTools.R
\name{Item parameter plot for mixed Rasch model}
\alias{Item parameter plot for mixed Rasch model}
\alias{mixRasch.plot}
\title{Item parameter plot for mixed Rasch model}
\usage{
mixRasch.plot(x, xlab, ylab)
}
\arguments{
\item{x}{A fitted multiple class mixture Rasch model}

\item{xlab}{user defined label for x axis}

\item{ylab}{user defined label for y axis}
}
\description{
This function produces a plot of item parameters by latent class for Rasch calibrated mixture models.
}
\details{
mixRaschTools Functions -----------------------------------------------------------

This function provides an item parameter plot that can be used to compare the different item
parameters in mixture Rasch models that contain two or more latent classes.
}
\examples{
##Example multiple class mixture Rasch models included with mixRaschTools
data(threeclass_ex)
##Item Parameter Plot
mixRasch.plot(threeclass_ex)
}
\author{
Pamela S Trantham
}
