% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixRaschTools.R
\name{Average Ability Levels for Multiple Class Mixed Rasch Model}
\alias{Average Ability Levels for Multiple Class Mixed Rasch Model}
\alias{avg.theta}
\title{Function for producing mean theta values for each class in a mixture Rasch model.}
\usage{
avg.theta(modelname)
}
\arguments{
\item{modelname}{A fitted multiple class mixture Rasch model}
}
\description{
This function produces mean ability levels for each class in a mixture Rasch model.
}
\details{
mixRasch Average Theta -----------------------------------------------------------

This function produces a matrix containing the average theta values for each class
included in mixture Rasch model.
}
\examples{
Example multiple class mixture Rasch models included with mixRaschTools
data(threeclass_ex)
Default item paramater plot
avg.theta(threeclass_ex)
}
\author{
Pamela S Trantham
}

