% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixRaschTools.R
\name{Item parameter plot for mixed Rasch model}
\alias{Item parameter plot for mixed Rasch model}
\alias{mixRasch.plot}
\title{Function for producing plots of item parameters for multiple class models.}
\usage{
mixRasch.plot(modelname, itemSet, xlab, ylab, file)
}
\arguments{
\item{modelname}{A fitted multiple class mixture Rasch model}
}
\description{
This function produces item parameters by class for Rasch calibrated mixture models.
}
\details{
mixRaschTools Functions -----------------------------------------------------------

This function provides an item parameter plot that can be used to compare the different item
parameters in mixture Rasch models that contain more than one class.
}
\examples{
Example multiple class mixture Rasch models included with mixRaschTools
data(threeclass_ex)
Default item paramater plot
mixRasch.plot(threeclass_ex)
}
\author{
Pamela S Trantham
}

